+++ management-api-server/src/main/java/com/datastax/mgmtapi/CqlService.java
@@ -65,4 +65,35 @@ public class CqlService {
       throw e;
     }
   }
+
+  /**
+   * Used for NodeOpsProvider implementations that are synchronous and may take a while to complete.
+   * (example node drain). The implementation here uses the Java driver's execution profile
+   * mechanism with the bundled application.conf that sets the request timeout to 0 seconds, which
+   * effectively disables the driver request timeout. Use this method with caution!. See
+   * application.conf in the resources folder.
+   */
+  public ResultSet executeSlowCql(File dbUnixSocketFile, String query)
+      throws ConnectionClosedException {
+    CqlSession session = UnixSocketCQLAccess.get(dbUnixSocketFile).orElse(null);
+
+    if (session == null || session.isClosed()) {
+      throw new ConnectionClosedException("Internal connection to Cassandra closed");
+    }
+
+    // build a statement and use the "slow" driver execution profile
+    SimpleStatementBuilder ssb = new SimpleStatementBuilder(query).setExecutionProfileName("slow");
+
+    try {
+      return session.execute(ssb.build());
+    } catch (NoNodeAvailableException e) {
+      try {
+        session.close();
+      } catch (Throwable t) {
+        // close quietly
+      }
+
+      throw e;
+    }
+  }
 }